requirement.txt
git+https://github.com/erikwijmans/Pointnet2_PyTorch.git#egg=pointnet2&subdirectory=pointnet2_ops_lib
protobuf==3.19
easydict==1.9
numpy==1.22
pandas==1.1.5
pomegranate
pyquaternion==0.9.9
pytorch==1.11.0
pytorch-lightning==1.3.8
PyYAML==5.4.1
scipy==1.5.4
Shapely==1.7.1
tdqm==0.0.1
tensorboard==2.4.1
torchmetrics==0.4.1
tqdm==4.61.1
nuscenes-devkit==1.1.9